remove full helper of harmonic mean; refactor
